requirements.txt
discord.py
aiohttp
characterai
websockets